home 페이지 리팩터링
