+++ src/App.css
@@ -1,133 +1,3 @@
 body {
   margin: 0px;
 }
-
-.Nav {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  top: 0px;
-  position: fixed;
-
-  height: 100vh;
-  width: 180px;
-  opacity: 0.95;
-  justify-content: space-evenly;
-}
-
-main {
-  margin-left: 60px;
-}
-
-.grid-container {
-  width: 600px;
-  /* 반응형 고려해야 함 */
-  border-radius: 10px;
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr 1fr;
-  grid-template-rows: 30px 35px 20px; /* 3개의 행을 생성 */
-  grid-template-areas:
-    'board title title title'
-    'board content content content'
-    'tag tag tag comment';
-
-  padding: 5px;
-  box-shadow: 1.5px 2px 5px 1px rgba(78, 120, 97, 0.3);
-}
-.grid-container * {
-  cursor: pointer;
-}
-.title {
-  grid-area: title;
-}
-.tag {
-  grid-area: tag;
-}
-
-.comment {
-  grid-area: comment;
-  text-align: right;
-}
-.content {
-  grid-area: content;
-  font-size: 13px;
-}
-.board {
-  grid-area: board;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.grid-container + .grid-container {
-  margin-top: 10px;
-}
-.grid-container:hover {
-  box-shadow: 3px 3px 10px 2px rgba(78, 120, 97, 0.3);
-  animation: hoverIn 0.2s ease-in-out forwards;
-}
-@keyframes hoverIn {
-  from {
-    transform: translateY(0rem);
-  }
-  to {
-    transform: translateY(-0.2rem);
-    background-color: rgba(192, 192, 192, 0.7);
-  }
-}
-@keyframes hoverOut {
-  from {
-    transform: translateY(-0.2rem);
-  }
-  to {
-    transform: translateY(0rem);
-    background-color: white;
-  }
-}
-.grid-container:not(:hover) {
-  animation: hoverOut 0.2s ease-in-out forwards;
-  background-color: white;
-}
-
-p {
-  margin: 0;
-  padding: 0;
-}
-
-h3 {
-  font-size: 16px;
-  font-weight: bold;
-  margin: 0px;
-  padding: 0px;
-}
-.ellipsis {
-  font-size: 12px;
-  overflow: hidden;
-  white-space: normal;
-  text-overflow: ellipsis;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  word-break: keep-all;
-}
-.listContainer {
-  border: 1px solid black;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 10px;
-  width: 1000px;
-}
-.ScreenLayout {
-  display: grid;
-  grid-template-columns: 180px 1fr;
-  width: 100%;
-  height: 100%;
-}
-.navContainer {
-  background-color: white;
-  border-right: 3px solid rgb(222, 212, 212);
-}
-.mainContainer {
-  background-color: white;
-}
