+++ src/index.css
@@ -1,7 +1,7 @@
-#root {
-  height: 100vh;
-  width: 100vw;
-}
 * {
   font-family: 'Nanum Gothic', sans-serif;
+  box-sizing: border-box;
+}
+body {
+  margin: 0px;
 }
