+++ src/pages/Home.module.css
@@ -1,0 +1,83 @@
+.threadItem {
+  width: 600px;
+  /* 반응형 고려해야 함 */
+  border-radius: 10px;
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr 1fr;
+  grid-template-rows: 30px 35px 20px; /* 3개의 행을 생성 */
+  grid-template-areas:
+    'board title title title'
+    'board content content content'
+    'tag tag tag comment';
+
+  /* 애니메이션  */
+  transform: translateY(0rem);
+  box-shadow: 1px 1px 3px 1px rgba(78, 120, 97, 0.3);
+  transition: box-shadow 0.3s, transform 0.3s;
+
+  padding: 5px;
+}
+.threadItem > * {
+  cursor: pointer;
+}
+.title {
+  grid-area: title;
+}
+.tag {
+  grid-area: tag;
+}
+
+.comment {
+  grid-area: comment;
+  text-align: right;
+}
+.content {
+  grid-area: content;
+  font-size: 13px;
+}
+.board {
+  grid-area: board;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.threadItem + .threadItem {
+  margin-top: 14px;
+}
+
+.threadItem:hover {
+  box-shadow: 3px 3px 10px 2px rgba(78, 120, 97, 0.3);
+  background-color: rgb(246, 236, 236);
+  transform: translateY(-0.4rem);
+}
+
+p {
+  margin: 0;
+  padding: 0;
+}
+
+h3 {
+  font-size: 16px;
+  font-weight: bold;
+  margin: 0px;
+  padding: 0px;
+}
+.ellipsis {
+  font-size: 12px;
+  overflow: hidden;
+  white-space: normal;
+  text-overflow: ellipsis;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+  word-break: keep-all;
+}
+.home-threadItem-Container {
+  /* border: 1px solid black; */
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 10px;
+  width: 900px;
+}
